requirements.txt
anthropic>=0.49.0
ollama>=0.4.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
